refactor: restaurantTables 관리자 확인 추가

+++ src/restaurantTables/restaurantTables.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Delete, Get, Request, UseGuards, Param } from '@nestjs/common';
+import { Controller, Post, Delete, Get, Request, UseGuards, Param, UnauthorizedException } from '@nestjs/common';
 import { RestaurantTablesService } from './restaurantTables.service';
 import { RestaurantTable } from './entities/restaurantTables.entity';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
@@ -10,22 +10,34 @@ export class RestaurantTablesController {
 
   @Get()
   async getTables(@Request() req): Promise<RestaurantTable[]> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.restaurantTablesService.getTables(restaurantId);
+      return this.restaurantTablesService.getTables(restaurantId);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Post()
   async addTable(@Request() req): Promise<RestaurantTable> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.restaurantTablesService.addTable(restaurantId);
+      return this.restaurantTablesService.addTable(restaurantId);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Delete(':table_num')
   async deleteTable(@Request() req, @Param('table_num') tableNum: number): Promise<void> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.restaurantTablesService.deleteTable(restaurantId, tableNum);
+      return this.restaurantTablesService.deleteTable(restaurantId, tableNum);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
